+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The following changes have been completed, and will be included in the next tagg
 * Added `PackageErrorTable` and `ERROR_TABLE`
 * Added `NotAFilepathError`
 * Added `PathApi` to model NodeJS's `path` module
+* Added `resolveFilepath()` data transform
 
 ### Fixes